+++ Rokt.Widget/android/src/oldarch/java/com/rokt/reactnativesdk/RoktEmbeddedViewManager.kt
@@ -1,12 +1,8 @@
 package com.rokt.reactnativesdk
 
-import com.facebook.react.bridge.Arguments
-import com.facebook.react.bridge.ReactContext
 import com.facebook.react.common.MapBuilder
 import com.facebook.react.uimanager.ThemedReactContext
 import com.facebook.react.uimanager.ViewGroupManager
-import com.facebook.react.uimanager.events.RCTEventEmitter
-import com.rokt.roktsdk.RoktWidgetDimensionCallBack
 import com.rokt.roktsdk.Widget
 
 /**
@@ -20,58 +16,31 @@ import com.rokt.roktsdk.Widget
  * You may obtain a copy of the License at https://rokt.com/sdk-license-2-0/
  */
 class RoktEmbeddedViewManager : ViewGroupManager<Widget>() {
-    companion object {
-        const val REACT_CLASS = "RoktNativeWidget"
-    }
-    override fun getName(): String = REACT_CLASS
+
+    private val impl = RoktEmbeddedViewManagerImpl()
+
+    override fun getName(): String = impl.getName()
 
     override fun createViewInstance(reactContext: ThemedReactContext): Widget {
-        val widget = Widget(reactContext)
-        setUpWidgetListeners(widget)
-        return widget
+        return impl.createViewInstance(reactContext)
     }
 
     override fun getExportedCustomDirectEventTypeConstants(): Map<String, Any>? {
         return MapBuilder.builder<String, Any>()
             .put(
-                "onWidgetHeightChanged",
-                MapBuilder.of("registrationName", "onWidgetHeightChanged")
+                RoktEmbeddedViewManagerImpl.EVENT_HEIGHT_CHANGED,
+                MapBuilder.of("registrationName", RoktEmbeddedViewManagerImpl.EVENT_HEIGHT_CHANGED)
             )
             .put(
-                "onWidgetMarginChanged",
-                MapBuilder.of("registrationName", "onWidgetMarginChanged")
+                RoktEmbeddedViewManagerImpl.EVENT_MARGIN_CHANGED,
+                MapBuilder.of("registrationName", RoktEmbeddedViewManagerImpl.EVENT_MARGIN_CHANGED)
             )
             .build()
     }
 
     override fun needsCustomLayoutForChildren(): Boolean = false
 
-    private fun setUpWidgetListeners(widget: Widget) {
-        widget.registerDimensionListener(object : RoktWidgetDimensionCallBack {
-            override fun onHeightChanged(height: Int) {
-                changeHeight(widget.context as ReactContext, height, widget.id)
-            }
-
-            override fun onMarginChanged(start: Int, top: Int, end: Int, bottom: Int) {
-                changeMargin(widget.context as ReactContext, widget.id, start, top, end, bottom)
-            }
-        })
-    }
-
-    fun changeHeight(context: ReactContext, height: Int, id: Int) {
-        val event = Arguments.createMap()
-        event.putString("height", height.toString())
-        context.getJSModule(RCTEventEmitter::class.java)
-            .receiveEvent(id, "onWidgetHeightChanged", event)
-    }
-
-    fun changeMargin(context: ReactContext, id: Int, start: Int, top: Int, end: Int, bottom: Int) {
-        val event = Arguments.createMap()
-        event.putString("marginLeft", start.toString())
-        event.putString("marginTop", top.toString())
-        event.putString("marginRight", end.toString())
-        event.putString("marginBottom", bottom.toString())
-        context.getJSModule(RCTEventEmitter::class.java)
-            .receiveEvent(id, "onWidgetMarginChanged", event)
+    fun setPlaceholderName(view: Widget?, value: String?) {
+        impl.setPlaceholderName(view, value)
     }
 }
